Add unit tests for ProfileDataCollector and ProfileDataExporter classes (#380)
* Add initial mock class and test file for ProfileDataCollector

* Mock FindExperiment method

* Add test for FindExperiment and fix a bug in the function

* Add tests for AddData and AddWindow in ProfileDataCollector

* Add a skeleton unit test code for ProfileDataExporter

* Mock ConvertToJson

* Add subcase for ConvertToJson method

* Add subcase for OutputToFile method

* Add subcase for AddExperiment method

* Split tests into multiple test cases

* Set request rate as double in json

* Address feedback

+++ src/c++/perf_analyzer/profile_data_exporter.cc
@@ -1,4 +1,4 @@
-// Copyright 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// Copyright (c) 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions
@@ -97,7 +97,7 @@ ProfileDataExporter::AddExperiment(
     value.SetUint64(raw_experiment.mode.concurrency);
   } else {
     mode = rapidjson::StringRef("request_rate");
-    value.SetUint64(raw_experiment.mode.request_rate);
+    value.SetDouble(raw_experiment.mode.request_rate);
   }
   experiment.AddMember("mode", mode, document_.GetAllocator());
   experiment.AddMember("value", value, document_.GetAllocator());
